requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
ansicon==1.89.0
appdirs==1.4.4
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cryptography==42.0.5
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.0.2
flyingcircus==0.1.4.1
frozenlist==1.4.1
gdown==5.0.0
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gspread==5.12.4
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.2.1
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
Pillow==10.1.0
protobuf==4.25.1
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyDrive==1.3.1
PyGetWindow==0.0.9
pygsheets==2.0.6
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyodbc==5.0.1
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytk==0.0.2.1
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.2
setuptools-scm==8.0.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
tk==0.1.0
tqdm==4.66.1
twilio==9.0.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.42.0
yarl==1.9.4
